refactor sidebar
