+++ packages/ui/src/components/Sidebar/index.ts
@@ -1,0 +1,12 @@
+import Sidebar, { type SidebarProps as TocSidebarProps } from "./Sidebar";
+import Wrapper from "./Wrapper";
+import Content from "./Content";
+
+const TableOfContents = {
+  Sidebar,
+  Wrapper,
+  Content,
+};
+
+export default TableOfContents;
+export type { TocSidebarProps };
